% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_resgene_prevalence.R
\name{calculate_resgene_prevalence}
\alias{calculate_resgene_prevalence}
\title{Functie om de prevalentie van een resistentiegen te berekenen}
\usage{
calculate_resgene_prevalence(gene, data)
}
\arguments{
\item{gene}{Het resistentiegen waarvan de prevalentie moet worden berekend}

\item{data}{Het dataframe met de gegevens, waarin de kolom resfinder_genes bevat}
}
\value{
De prevalentie van het opgegeven resistentiegen in de dataset
}
\description{
Deze functie berekent de prevalentie van een opgegeven resistentiegen in de dataset.
}
