% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_antibiotic_class.R
\name{get_antibiotic_class}
\alias{get_antibiotic_class}
\title{Functie om de antibiotic class op te halen op basis van postcode en geboortejaar}
\usage{
get_antibiotic_class(zipcode, year_of_birth, data)
}
\arguments{
\item{zipcode}{De postcode van de persoon}

\item{year_of_birth}{Het geboortejaar van de persoon}

\item{data}{Het dataframe met de gegevens waarin de kolommen zipcode, antibiotic_class en person_birthyear bevat.}
}
\value{
De antibiotic class van de persoon, of een foutmelding als er geen gegevens zijn gevonden
}
\description{
Deze functie zoekt in de gegeven dataset naar de antibiotic class van een persoon met de opgegeven postcode en geboortejaar.
}
