% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_info.R
\name{get_date_info}
\alias{get_date_info}
\title{Functie om informatie te verkrijgen voor een specifieke bemonsteringsdatum}
\usage{
get_date_info(data, sampling_date)
}
\arguments{
\item{data}{Het dataframe met de gegevens, waarin de kolommen sampling_date, institute_city, institute_province, person_gender bevat.}

\item{sampling_date}{De specifieke bemonsteringsdatum waarvoor de informatie moet worden verkregen}
}
\value{
Een dataframe met de unieke combinatie van stad, provincie en geslacht voor de opgegeven bemonsteringsdatum
}
\description{
Deze functie filtert de dataset voor een specifieke bemonsteringsdatum en retourneert de unieke combinatie van stad, provincie en geslacht.
}
