% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_metadata.R
\name{Load_metadata}
\alias{Load_metadata}
\title{Functie om metadata te laden}
\usage{
Load_metadata(file_path)
}
\arguments{
\item{file_path}{Het pad naar het CSV-bestand}
}
\value{
Een dataframe met de geladen metadata
}
\description{
Deze functie leest een CSV-bestand in vanuit het opgegeven pad en retourneert de metadata.
}
