% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ab_resis.R
\name{plot_ab_resis}
\alias{plot_ab_resis}
\title{Functie om het aantal antibioticaresistentie classes per provincie te plotten}
\usage{
plot_ab_resis(data)
}
\arguments{
\item{data}{Een dataframe met de gegevens, waarin de kolom institute_province de provincie aanduidt en de kolom n_classes_antibiotica het aantal antibioticaresistentie classes bevat.}
}
\value{
Een ggplot-object met de barplot van het aantal antibioticaresistentie classes per provincie.
}
\description{
Deze functie groepeert de gegevens per provincie en telt het aantal antibioticaresistentie classes per provincie. Vervolgens maakt het een barplot van het aantal classes per provincie.
}
